Silence direnv output
When entering a directory with a Nix flake, direnv automatically sets up
the env vars for the directory using the contents of the flake. By
default, direnv logs a diff of all the env vars that have been changed,
but this gets verbose quickly with Nix, so we avoid doing so.